Got rid of unnecessary unboxing

+++ core/schema/src/main/java/org/opennms/core/schema/Migrator.java
@@ -217,7 +217,7 @@ public class Migrator {
                                              "Unsupported database version \"%f\" -- you need at least %f and less than %f.  "
                                                      + "Use the \"-Q\" option to disable this check if you feel brave and are willing "
                                                      + "to find and fix bugs found yourself.",
-                                                     dbv.floatValue(), POSTGRES_MIN_VERSION, POSTGRES_MAX_VERSION_PLUS_ONE
+                                                     dbv, POSTGRES_MIN_VERSION, POSTGRES_MAX_VERSION_PLUS_ONE
                 );
 
         if (dbv < POSTGRES_MIN_VERSION || dbv >= POSTGRES_MAX_VERSION_PLUS_ONE) {
